refactor: alteração da model cliente

+++ src/database/models/cliente.js
@@ -3,7 +3,7 @@ const clienteSchema = (sequelize, DataTypes) => {
   const userTable = sequelize.define(
     "Cliente",
     {
-      id: { autoIncrement: true, primaryKey: true, type: DataTypes.INTEGER(5) },
+      id: { autoIncrement: true, primaryKey: true, type: DataTypes.INTEGER },
       codCliente: DataTypes.STRING,
       password: DataTypes.STRING,
          },
